# proto/openapi.yaml
openapi: 3.1.0
info:
  title: Nakvali API
  version: 0.1.0
  description: >
    Contract-first spec for the Nakvali backend. Skeleton — grows with each phase.
servers:
  - url: /api/v1
security:
  - privateAlphaAccessKey: []
paths:
  /me:
    get:
      summary: Current user profile
      operationId: getMe
      security:
        - privateAlphaAccessKey: []
          firebaseBearer: []
      responses:
        "200":
          description: Verified Firebase identity mapped to the local Nakvali user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401":
          description: Missing, expired or invalid Firebase ID token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Firebase verification or the product database is unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /activities:
    post:
      summary: Create an activity through the opt-in legacy raw-ingestion path
      description: Available only when the server enables `RAW_UPLOADS_ENABLED` for development.
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateActivityRequest" }
      responses:
        "201":
          description: Activity created in status `recording`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateActivityResponse" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Database unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /activities/{id}/raw:
    put:
      summary: Upload the raw recording (gzipped JSON Lines, see raw-recording-format.md)
      description: Development-only compatibility route; production keeps raw on-device.
      operationId: uploadRawRecording
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/gzip:
            schema: { type: string, format: binary }
      responses:
        "204":
          description: Stored to object storage; activity status → `raw_uploaded`
        "400":
          description: Invalid input (e.g. empty body)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Unknown activity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: Body exceeds 256 MB limit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: Content-Type is not application/gzip
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Database unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /activities/{id}/finish:
    post:
      summary: Finish an activity (status → `uploaded`)
      description: Development-only companion to the opt-in raw-ingestion path.
      operationId: finishActivity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FinishActivityRequest" }
      responses:
        "204":
          description: Activity finished
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Unknown activity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Database unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /strava/connect:
    post:
      summary: Start or restart the one-time Strava OAuth connection
      operationId: beginStravaConnect
      security:
        - privateAlphaAccessKey: []
          deviceBearer: []
      responses:
        "200":
          description: Mobile OAuth URL to open outside the app
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StravaConnectResponse" }
        "401":
          description: Missing or invalid anonymous device credential
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Strava broker is not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /strava/connection:
    get:
      summary: Get this app installation's Strava connection state
      operationId: getStravaConnection
      security:
        - privateAlphaAccessKey: []
          deviceBearer: []
      responses:
        "200":
          description: Current connection state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StravaConnection" }
        "401":
          description: Missing or invalid anonymous device credential
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /strava/oauth/callback:
    get:
      summary: Strava browser callback; exchanges code server-side and redirects to Nakvali
      operationId: completeStravaConnect
      security: []
      parameters:
        - { name: state, in: query, schema: { type: string } }
        - { name: code, in: query, schema: { type: string } }
        - { name: scope, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
      responses:
        "302":
          description: Redirect to `nakvali://strava/connected?result=...`
  /strava/exports:
    post:
      summary: Idempotently export one canonical processed GPX to Strava
      description: >
        Raw GPS, IMU, barometer and health data are never accepted. Repeating the
        same external_id polls the existing Strava upload instead of submitting
        a second activity.
      operationId: exportActivityToStrava
      security:
        - privateAlphaAccessKey: []
          deviceBearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, external_id, title, sport_type]
              properties:
                file: { type: string, format: binary, description: Canonical processed GPX }
                external_id: { type: string, maxLength: 200 }
                title: { type: string, maxLength: 200 }
                description: { type: string, maxLength: 5000 }
                sport_type:
                  type: string
                  enum: [MountainBikeRide, EMountainBikeRide]
      responses:
        "200":
          description: Strava activity is ready
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StravaExportStatus" }
        "202":
          description: Upload accepted and still processing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StravaExportStatus" }
        "409":
          description: Strava has not been connected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: GPX exceeds 32 MB
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Strava rejected or could not process the GPX
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StravaExportStatus" }
components:
  securitySchemes:
    privateAlphaAccessKey:
      type: apiKey
      in: header
      name: X-Nakvali-Access-Key
      description: >
        Shared private-alpha perimeter key compiled only into owner test builds.
        This is not user authentication and must be replaced before public distribution.
    deviceBearer:
      type: http
      scheme: bearer
      description: >
        Random per-installation credential generated and retained by Android.
        The backend stores only its SHA-256 hash; it is not a Nakvali user account.
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Firebase ID token obtained by the signed-in Android user and verified by the
        API with the Firebase Admin SDK. Its verified uid is the external identity.
  schemas:
    UserProfile:
      type: object
      required: [id, email, display_name, avatar_url, email_verified]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        display_name: { type: string }
        avatar_url: { type: string, format: uri }
        email_verified: { type: boolean }
    CreateActivityRequest:
      type: object
      required: [sport, started_at_ms]
      properties:
        sport: { type: string, examples: [downhill] }
        started_at_ms: { type: integer, format: int64, description: Unix epoch milliseconds }
    CreateActivityResponse:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
    FinishActivityRequest:
      type: object
      required: [ended_at_ms]
      properties:
        ended_at_ms: { type: integer, format: int64, description: Unix epoch milliseconds }
        title:
          type: string
          maxLength: 200
          description: User-entered title. Omitted or empty means unset (stored as NULL).
        description:
          type: string
          maxLength: 5000
          description: User-entered description. Omitted or empty means unset (stored as NULL).
        bike:
          type: string
          maxLength: 100
          description: User-entered bike name. Omitted or empty means unset (stored as NULL).
        bike_type:
          type: string
          enum: [full_sus, hardtail, ebike, other]
          description: Bike category. Omitted or empty means unset (stored as NULL).
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, examples: [db_unavailable] }
    StravaConnectResponse:
      type: object
      required: [authorize_url]
      properties:
        authorize_url: { type: string, format: uri }
    StravaConnection:
      type: object
      required: [connected]
      properties:
        connected: { type: boolean }
        athlete_id: { type: integer, format: int64 }
        athlete_name: { type: string }
    StravaExportStatus:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [processing, uploaded, failed] }
        strava_upload_id: { type: integer, format: int64 }
        strava_activity_id: { type: integer, format: int64 }
        error: { type: string }
    GpsPoint:
      type: object
      required: [timestampMs, lat, lon]
      properties:
        timestampMs: { type: integer, format: int64 }
        lat: { type: number, format: double }
        lon: { type: number, format: double }
        altitudeM: { type: number, format: double }
        accuracyM: { type: number, format: float }
        speedMps: { type: number, format: float }
        bearingDeg: { type: number, format: float }